Restructured CAN library
