+++ libs/CAN/include/CAN/messages/science.h
@@ -1,0 +1,56 @@
+#pragma once
+
+#include "science_codec.h"
+#include "CAN/codec.h"
+
+
+#include <cstdint>
+
+static constexpr uint8_t ENABLE_SCIENCE_DLC = 1;
+static constexpr uint8_t MOVE_AUGER_DLC = 8;
+static constexpr uint8_t HOME_AUGER_DLC = 0;
+static constexpr uint8_t ENABLE_DRILL_DLC = 1;
+static constexpr uint8_t MOVE_SLIDE_DLC = 1;
+
+struct EnableScienceMsg {
+    uint8_t enable;
+};
+
+enum class ScienceState : uint8_t {
+    Active = 0,
+    Stop = 1,
+};
+
+enum class ScienceDir : uint8_t {
+    Reverse = 0,
+    Forward = 1,
+};
+
+struct MoveAugerMsg {
+    ScienceState state;
+    ScienceDir direction;
+};
+
+struct HomeAugerMsg {
+
+};
+
+struct EnableDrillMsg {
+    uint8_t enable;
+};
+
+// XXX: so ik this needs to be defined for the CAN library so ROS can have the
+// types, but i have fears about desync between this and the teensy science code
+// so i wonder if theres a way for them to share the enum.
+// For now just going to update them identically
+enum class SlidePosition : uint8_t {
+    Home = 0x00,
+    OpenSample = 0x01,
+    SampleCache = 0x02,
+    MiddleContainer = 0x03,
+    EndSample = 0x04,
+};
+
+struct MoveSlideMsg {
+    SlidePosition position;
+};
