+++ libs/CAN/tools/codegen/templates/codec_template.j2
@@ -6,6 +6,8 @@
 #include "{{ DLC_HEADER }}"      // DLC constants
 {% endif %}
 
+#include "CAN/codec.h"
+
 // ---------------- Auto-generated encode/decode ----------------
 
 {% for msg in MESSAGES %}
